+++ internal/controller/cluster_controller.go
@@ -156,6 +156,7 @@ func SetupClusterMonitoringReconciler(mgr manager.Manager, cfg config.Config, lo
 		OrganizationRepository:  organizationRepository,
 		TenantRepository:        tenantRepository,
 		LogsAuthManager:         lokiAuthManager,
+		MetricsAuthManager:      mimirAuthManager,
 		TracesAuthManager:       tempoAuthManager,
 	}
 
@@ -381,8 +382,9 @@ func (r *ClusterMonitoringReconciler) reconcileAlloyServices(ctx context.Context
 		}
 	}
 
-	// alloy-event specific: Alloy events configuration - deployment that handles both kube event logs and traces - TODO rename alloy-events to alloy-cluster
-	if r.Config.Logging.IsLoggingEnabled(cluster) || r.Config.Tracing.IsTracingEnabled(cluster) {
+	// alloy-event specific: Alloy events configuration - deployment that handles kube event logs, traces, and OTLP ingestion - TODO rename alloy-events to alloy-cluster
+	otlpEnabled := (r.Config.Monitoring.OTLPEnabled && r.Config.Monitoring.IsMonitoringEnabled(cluster)) || (r.Config.Logging.OTLPEnabled && r.Config.Logging.IsLoggingEnabled(cluster))
+	if r.Config.Logging.IsLoggingEnabled(cluster) || r.Config.Tracing.IsTracingEnabled(cluster) || otlpEnabled {
 		// Create or update Alloy events configuration
 		err = r.AlloyEventsService.ReconcileCreate(ctx, cluster, observabilityBundleVersion)
 		if err != nil {